# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'API DOCS BOT|POS'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/produkKategori/list:
    post:
      summary: 'Produk Kategori List'
      operationId: produkKategoriList
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: ''
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - 'Produk Kategori'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start:
                  type: integer
                  description: 'start data.'
                  example: 0
                limit:
                  type: integer
                  description: 'limit data.'
                  example: 10
                filter:
                  type: string
                  description: 'Text biasa bisa diisi bisa tidak.'
                  example: null
              required:
                - start
                - limit
                - filter
...
  /api/produkKategori/create:
    post:
      summary: 'Produk Kategori Create'
      operationId: produkKategoriCreate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 200
                    message: 'Request Berhasil.'
                  data: {  }
                  request:
                    xxx_id: 1
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 200
                      message:
                        type: string
                        example: 'Request Berhasil.'
                  data:
                    type: object
                    properties: {  }
                  request:
                    type: object
                    properties:
                      xxx_id:
                        type: integer
                        example: 1
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - 'Produk Kategori'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ProdukKategori_kode:
                  type: string
                  description: ProdukKategori_kode.
                  example: null
                ProdukKategori_nama:
                  type: string
                  description: ProdukKategori_nama.
                  example: null
  /api/produkKategori/update:
    put:
      summary: 'Produk Kategori Update'
      operationId: produkKategoriUpdate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 200
                    message: 'Request Berhasil.'
                  data: {  }
                  request:
                    xxx_id: 1
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 200
                      message:
                        type: string
                        example: 'Request Berhasil.'
                  data:
                    type: object
                    properties: {  }
                  request:
                    type: object
                    properties:
                      xxx_id:
                        type: integer
                        example: 1
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - 'Produk Kategori'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kategori_id:
                  type: string
                  description: kategori_id.
                  example: '1'
                ProdukKategori_kode:
                  type: string
                  description: ProdukKategori_kode.
                  example: null
                ProdukKategori_nama:
                  type: string
                  description: ProdukKategori_nama.
                  example: null
              required:
                - kategori_id
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []